DiaBot/requirements.txt
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.3
colorama==0.4.6
cryptography==41.0.2
Flask==2.3.2
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
frozenlist==1.4.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
mysql-connector==2.2.9
mysql-connector-python==8.0.33
protobuf==3.20.3
pycparser==2.21
PyJWT==2.7.0
pyodbc==4.0.39
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
twilio==8.5.0
urllib3==2.0.4
Werkzeug==2.3.6
yarl==1.9.2
